requirements.txt
bson
pykafka
avro-python3